fix compiler warning TSL2591

+++ lib/lib_i2c/Adafruit_TSL2591_Library/Adafruit_TSL2591.cpp
@@ -214,7 +214,7 @@ void Adafruit_TSL2591::setGain(tsl2591Gain_t gain)
 
   enable();
   _gain = gain;
-  write8(TSL2591_COMMAND_BIT | TSL2591_REGISTER_CONTROL, _integration | _gain);
+  write8(TSL2591_COMMAND_BIT | TSL2591_REGISTER_CONTROL, static_cast<uint8_t>(_integration) | static_cast<uint8_t>(_gain));
   disable();
 }
